+++ view/src/main/java/com/media2359/euphoria/view/client/project/ProjectProperties.java
@@ -12,7 +12,7 @@ public interface ProjectProperties extends PropertyAccess<ProjectDTO> {
 
 	ValueProvider<ProjectDTO, String> name();
 
-	ValueProvider<ProjectDTO, Integer> manDaysLeft();
+	ValueProvider<ProjectDTO, Double> manDaysLeft();
 
 	ValueProvider<ProjectDTO, Integer> milestoneCount();
